requirements.txt
pydantic[email]~=2.9.2
pytest~=8.3.2
python-dateutil~=2.9.0.post0
fastapi==0.114.2
uvicorn==0.30.6
httpx==0.27.2
sqlalchemy==2.0.36
python-dotenv==1.0.1
pymysql==1.1.1
beautifulsoup4==4.12.3
pyap==0.3.1
price-parser==0.3.4
pycountry==24.6.1
pandas==2.2.3
numpy==2.1.3
numpy-financial==1.0.0
requests==2.32.3
openai==1.55.1